experiment 1
my new branch experimenting with enemy navigation, fraught with
concerns that I don’t understand this git shit
